move tests from exp_test file to message_test
